aws/requirements.txt
aws-cdk-lib==2.106.1
constructs>=10.0.0,<11.0.0
requests
